Move `patch_subcript()` to the dict-common directory.
Multiple backends use this routine; place it in the common area.
