+++ link-grammar/dict-common/dict-utils.c
@@ -18,7 +18,30 @@
 #include "connectors.h"
 #include "dict-api.h"
 #include "string-set.h"
+#include "dict-defines.h" // for SUBSCRIPT_MARK
 #include "dict-utils.h"
+
+
+/** Replace the right-most dot with SUBSCRIPT_MARK */
+void patch_subscript(char * s)
+{
+	char *ds, *de;
+	int dp;
+	ds = strrchr(s, SUBSCRIPT_DOT);
+	if (!ds) return;
+
+	/* A dot at the end or a dot followed by a number is NOT
+	 * considered a subscript */
+	de = ds + 1;
+	if (*de == '\0') return;
+	dp = (int) *de;
+
+	/* If it's followed by a UTF8 char, its NOT a subscript */
+	if (127 < dp || dp < 0) return;
+	/* assert ((0 < dp) && (dp <= 127), "Bad dictionary entry!"); */
+	if (isdigit(dp)) return;
+	*ds = SUBSCRIPT_MARK;
+}
 
 /* ======================================================== */
 /* Public API ... */
